buildkite step analysis: move and rename
